double-clicking a section will close it and room can be minimized and order can be set in a shelf through backend

+++ src/main/java/com/project/jpa/controller/BookController.java
@@ -84,7 +84,7 @@ public class BookController {
 	// New Mappings
 	
 	@GetMapping("/books/{book}")
-	public List<Book> booksOfShelf(@PathVariable int book) {
+	public List<Book> booksOfShelf(@PathVariable long book) {
 		try {
 			return shelfRepo.findById(book).orElseThrow().getBooks();
 		}
@@ -95,7 +95,7 @@ public class BookController {
 	
 
 	@PostMapping("/books/{shelfId}/users/{userId}")
-	public List<Book> bookForShelf(@PathVariable int shelfId, @PathVariable String userId, @RequestBody List<Book> books) {
+	public List<Book> bookForShelf(@PathVariable long shelfId, @PathVariable String userId, @RequestBody List<Book> books) {
 //		Room room = roomRepo.findById(shelf.getBookcase().getRoom().getId()).orElseThrow();
 		Shelf shelf = shelfRepo.findById(shelfId).orElseThrow();
 		User user = userRepo.findByUsername(userId).get(0);		
@@ -119,7 +119,7 @@ public class BookController {
 	}
 	
 	@DeleteMapping("/books/{bookId}/shelves/{shelfId}/users/{userId}")
-	public Map<String, Boolean> bookFromShelf(@PathVariable Integer bookId, @PathVariable Integer shelfId, @PathVariable String userId) {
+	public Map<String, Boolean> bookFromShelf(@PathVariable Long bookId, @PathVariable Long shelfId, @PathVariable String userId) {
 		Map<String, Boolean> res = new HashMap<>();
 		try {
 			User user = userRepo.findByUsername(userId).get(0);
@@ -138,7 +138,7 @@ public class BookController {
 
 
 	@PutMapping("/books/{bookId}")
-	public Book updateBookForShelf(@PathVariable int bookId, @RequestBody Book newBook) {
+	public Book updateBookForShelf(@PathVariable long bookId, @RequestBody Book newBook) {
 		
 		Book oldBook = bookRepo.findById(bookId).orElseThrow();
 		
@@ -210,10 +210,10 @@ public class BookController {
 	}
 	
 	@GetMapping("/books/{id}/coord")
-	public Map<String, Integer> getBookCoordinates(@PathVariable int id) {
+	public Map<String, Long> getBookCoordinates(@PathVariable long id) {
 		System.out.println("search by all");
 		Book book = bookRepo.getById(id);
-		Map<String, Integer> res = new HashMap<>();
+		Map<String, Long> res = new HashMap<>();
 		res.put("book", book.getId());
 		res.put("shelf", book.getShelf().getId());
 		res.put("bookcase", book.getShelf().getBookcase().getId());
@@ -224,55 +224,55 @@ public class BookController {
 	// Search for books in room
 	
 	@GetMapping("/books/search/in/room/{id}/title={title}")
-	public List<Book> getByTitleInRoom(@PathVariable int id, @PathVariable String title) {
+	public List<Book> getByTitleInRoom(@PathVariable long id, @PathVariable String title) {
 		System.out.println("search by title in room");
 		return bookRepo.findTitleInRoom(title, id);
 	}	
 	
 	@GetMapping("/books/search/in/room/{id}/genre={genre}")
-	public List<Book> getByGenreInRoom(@PathVariable int id, @PathVariable String genre) {
+	public List<Book> getByGenreInRoom(@PathVariable long id, @PathVariable String genre) {
 		System.out.println("search by title in room");
 		return bookRepo.findGenreInRoom(genre, id);
 	}	
 	
 	@GetMapping("/books/search/in/room/{id}/genretitle={search}")
-	public List<Book> getByGenreAndTitleInRoom(@PathVariable int id, @PathVariable String search) {
+	public List<Book> getByGenreAndTitleInRoom(@PathVariable long id, @PathVariable String search) {
 		System.out.println("search by genre and title in room");
 		return bookRepo.findGenreAndTitleInRoom(search, id);
 	}	
 	
 	@GetMapping("/books/search/in/room/{id}/all={all}")
-	public List<Book> getByAllInRoom(@PathVariable int id, @PathVariable String all) {
+	public List<Book> getByAllInRoom(@PathVariable long id, @PathVariable String all) {
 		System.out.println("search by more in room");
 		return bookRepo.findAllInRoom(all, id);
 	}	
 	
 	@GetMapping("/books/search/in/room/{id}/color={color}")
-	public List<Book> getByColorInRoom(@PathVariable int id, @PathVariable String color) {
+	public List<Book> getByColorInRoom(@PathVariable long id, @PathVariable String color) {
 		System.out.println("search by color in room");
 		return bookRepo.findColorInRoom(color, id);
 	}
 	
 	@GetMapping("/books/search/in/room/{id}/author={author}")
-	public List<Book> getByAuthorInRoom(@PathVariable int id, @PathVariable String author) {
+	public List<Book> getByAuthorInRoom(@PathVariable long id, @PathVariable String author) {
 		System.out.println("search by author in room");
 		return bookRepo.findAuthorInRoom(author, id);
 	}
 	
 	@GetMapping("/books/search/in/room/{id}/more={more}")
-	public List<Book> getByMoreInRoom(@PathVariable int id, @PathVariable String more) {
+	public List<Book> getByMoreInRoom(@PathVariable long id, @PathVariable String more) {
 		System.out.println("search by more in room");
 		return bookRepo.findMoreInRoom(more, id);
 	}
 	
 	@GetMapping("/books/search/in/room/{id}/published={low}&{high}")
-	public List<Book> getByPublishDateInRoom(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPublishDateInRoom(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by publish date in room");
 		return bookRepo.findPublishDateInRoom(low, high, id);
 	}
 	
 	@GetMapping("/books/search/in/room/{id}/pages={low}&{high}")
-	public List<Book> getByPagesInRoom(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPagesInRoom(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by pages in room");
 		return bookRepo.findPagesInRoom(low, high, id);
 	}
@@ -281,55 +281,55 @@ public class BookController {
 	// Search for books in bookcase
 	
 	@GetMapping("/books/search/in/bookcase/{id}/title={title}")
-	public List<Book> getByTitleInBookcase(@PathVariable int id, @PathVariable String title) {
+	public List<Book> getByTitleInBookcase(@PathVariable long id, @PathVariable String title) {
 		System.out.println("search by title in bookcase");
 		return bookRepo.findTitleInBookcase(title, id);
 	}	
 	
 	@GetMapping("/books/search/in/bookcase/{id}/genre={genre}")
-	public List<Book> getByGenreInBookcase(@PathVariable int id, @PathVariable String genre) {
+	public List<Book> getByGenreInBookcase(@PathVariable long id, @PathVariable String genre) {
 		System.out.println("search by genre in bookcase");
 		return bookRepo.findGenreInBookcase(genre, id);
 	}	
 	
 	@GetMapping("/books/search/in/bookcase/{id}/genretitle={search}")
-	public List<Book> getByGenreAndTitleInBookcase(@PathVariable int id, @PathVariable String search) {
+	public List<Book> getByGenreAndTitleInBookcase(@PathVariable long id, @PathVariable String search) {
 		System.out.println("search by genre and title in bookcase");
 		return bookRepo.findGenreAndTitleInBookcase(search, id);
 	}	
 	
 	@GetMapping("/books/search/in/bookcase/{id}/all={all}")
-	public List<Book> getByAllInBookcase(@PathVariable int id, @PathVariable String all) {
+	public List<Book> getByAllInBookcase(@PathVariable long id, @PathVariable String all) {
 		System.out.println("search by all in bookcase");
 		return bookRepo.findAllInBookcase(all, id);
 	}	
 	
 	@GetMapping("/books/search/in/bookcase/{id}/color={color}")
-	public List<Book> getByColorInBookcase(@PathVariable int id, @PathVariable String color) {
+	public List<Book> getByColorInBookcase(@PathVariable long id, @PathVariable String color) {
 		System.out.println("search by color in bookcase");
 		return bookRepo.findColorInBookcase(color, id);
 	}
 	
 	@GetMapping("/books/search/in/bookcase/{id}/author={author}")
-	public List<Book> getByAuthorInBookcase(@PathVariable int id, @PathVariable String author) {
+	public List<Book> getByAuthorInBookcase(@PathVariable long id, @PathVariable String author) {
 		System.out.println("search by author in bookcase");
 		return bookRepo.findAuthorInBookcase(author, id);
 	}
 	
 	@GetMapping("/books/search/in/bookcase/{id}/more={more}")
-	public List<Book> getByMoreInBookcase(@PathVariable int id, @PathVariable String more) {
+	public List<Book> getByMoreInBookcase(@PathVariable long id, @PathVariable String more) {
 		System.out.println("search by more in bookcase");
 		return bookRepo.findMoreInBookcase(more, id);
 	}
 	
 	@GetMapping("/books/search/in/bookcase/{id}/published={low}&{high}")
-	public List<Book> getByPublishDateInBookcase(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPublishDateInBookcase(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by publish date in bookcase");
 		return bookRepo.findPublishDateInBookcase(low, high, id);
 	}
 	
 	@GetMapping("/books/search/in/bookcase/{id}/pages={low}&{high}")
-	public List<Book> getByPagesInBookcase(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPagesInBookcase(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by pages in bookcase");
 		return bookRepo.findPagesInBookcase(low, high, id);
 	}
@@ -338,55 +338,55 @@ public class BookController {
 	// Search for books in shelf
 	
 	@GetMapping("/books/search/in/shelf/{id}/title={title}")
-	public List<Book> getByTitleInShelf(@PathVariable int id, @PathVariable String title) {
+	public List<Book> getByTitleInShelf(@PathVariable long id, @PathVariable String title) {
 		System.out.println("search by title in shelf");
 		return bookRepo.findTitleInShelf(title, id);
 	}	
 	
 	@GetMapping("/books/search/in/shelf/{id}/genre={genre}")
-	public List<Book> getByGenreInShelf(@PathVariable int id, @PathVariable String genre) {
+	public List<Book> getByGenreInShelf(@PathVariable long id, @PathVariable String genre) {
 		System.out.println("search by genre in shelf");
 		return bookRepo.findGenreInShelf(genre, id);
 	}	
 	
 	@GetMapping("/books/search/in/shelf/{id}/genretitle={search}")
-	public List<Book> getByGenreAndTitleInShelf(@PathVariable int id, @PathVariable String search) {
+	public List<Book> getByGenreAndTitleInShelf(@PathVariable long id, @PathVariable String search) {
 		System.out.println("search by genre and title in shelf");
 		return bookRepo.findGenreAndTitleInShelf(search, id);
 	}	
 	
 	@GetMapping("/books/search/in/shelf/{id}/all={all}")
-	public List<Book> getByAllInShelf(@PathVariable int id, @PathVariable String all) {
+	public List<Book> getByAllInShelf(@PathVariable long id, @PathVariable String all) {
 		System.out.println("search by all in shelf");
 		return bookRepo.findAllInShelf(all, id);
 	}	
 	
 	@GetMapping("/books/search/in/shelf/{id}/color={color}")
-	public List<Book> getByColorInShelf(@PathVariable int id, @PathVariable String color) {
+	public List<Book> getByColorInShelf(@PathVariable long id, @PathVariable String color) {
 		System.out.println("search by color in shelf");
 		return bookRepo.findColorInShelf(color, id);
 	}
 	
 	@GetMapping("/books/search/in/shelf/{id}/author={author}")
-	public List<Book> getByAuthorInShelf(@PathVariable int id, @PathVariable String author) {
+	public List<Book> getByAuthorInShelf(@PathVariable long id, @PathVariable String author) {
 		System.out.println("search by author in shelf");
 		return bookRepo.findAuthorInShelf(author, id);
 	}
 	
 	@GetMapping("/books/search/in/shelf/{id}/more={more}")
-	public List<Book> getByMoreInShelf(@PathVariable int id, @PathVariable String more) {
+	public List<Book> getByMoreInShelf(@PathVariable long id, @PathVariable String more) {
 		System.out.println("search by more in shelf");
 		return bookRepo.findMoreInShelf(more, id);
 	}
 	
 	@GetMapping("/books/search/in/shelf/{id}/published={low}&{high}")
-	public List<Book> getByPublishDateInShelf(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPublishDateInShelf(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by publish date in shelf");
 		return bookRepo.findPublishDateInShelf(low, high, id);
 	}
 	
 	@GetMapping("/books/search/in/shelf/{id}/pages={low}&{high}")
-	public List<Book> getByPagesInShelf(@PathVariable int id, @PathVariable int low, @PathVariable int high) {
+	public List<Book> getByPagesInShelf(@PathVariable long id, @PathVariable int low, @PathVariable int high) {
 		System.out.println("search by pages in shelf");
 		return bookRepo.findPagesInShelf(low, high, id);
 	}
