+++ src/main/java/com/project/jpa/repository/BookcaseRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 import com.project.jpa.model.Bookcase;
 import com.project.jpa.model.User;
 
-public interface BookcaseRepository extends JpaRepository<Bookcase,Integer>{
+public interface BookcaseRepository extends JpaRepository<Bookcase,Long>{
 
 	@Query("select r from Room r JOIN r.bookcases k")
 	List<User> joinRoomAndBookcases(@Param("n") String username);
