+++ src/main/java/com/project/jpa/repository/ShelfRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.project.jpa.model.Shelf;
 
-public interface ShelfRepository extends JpaRepository<Shelf,Integer> {
+public interface ShelfRepository extends JpaRepository<Shelf,Long> {
